✨ Feat: 토스 결제 승인 요청 성공 시, 토스에서 제공하는 paymentKey, approvedAt, status 저장 추가 구현

+++ src/main/java/com/ssg/wannavapibackend/dto/request/TossPaymentRequestDTO.java
@@ -21,7 +21,7 @@ public class TossPaymentRequestDTO {
      */
     private String paymentKey;  // 고유 결제 키
     private String orderId;     // 주문 번호
-    private BigDecimal amount;  // 결제 금액
+    private Double amount;  // 결제 금액
 
     public String toJson() throws JsonProcessingException {
         ObjectMapper objectMapper = new ObjectMapper();
